v1.4.1

+++ my_os_v1.4/userprog/tss.c
@@ -53,7 +53,7 @@ void update_tss_esp(struct task_struct* pthread)
 }
 
 /* 创建gdt描述符 */
-static struct gdt_desc make_gdt_desc(uint32_t* desc_addr,uint32_t limit,uint32_t attr_low,uint8_t attr_high)
+static struct gdt_desc make_gdt_desc(uint32_t* desc_addr,uint32_t limit,uint8_t attr_low,uint8_t attr_high)
 {
     uint32_t desc_base = (uint32_t)desc_addr;
     struct gdt_desc desc;
